Cast to float explicitly.

+++ src/execution/vm/bytecode_generator.cpp
@@ -3288,7 +3288,7 @@ void BytecodeGenerator::VisitLitExpr(ast::LitExpr *node) {
     }
     case ast::LitExpr::LitKind::Float: {
       if (const auto size = node->GetType()->GetSize(); size == 4) {
-        GetEmitter()->EmitAssignImm4F(target, node->Float64Val());
+        GetEmitter()->EmitAssignImm4F(target, static_cast<float>(node->Float64Val()));
       } else {
         NOISEPAGE_ASSERT(size == 8, "Invalid float literal size. Must be 4-, or 8-bytes.");
         GetEmitter()->EmitAssignImm8F(target, node->Float64Val());
